% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F3_densify_score.R
\name{densify_score}
\alias{densify_score}
\title{densify_score}
\usage{
densify_score(
  documentation,
  exponent_prop_coded_data = 1,
  exponent_available_data_points = 1,
  exponent_lowest_language_score = 1,
  exponent_taxonomic_diversity = 1
)
}
\arguments{
\item{documentation}{A data frame containing the output documentation from the \code{densify_steps} function.}

\item{exponent_prop_coded_data}{Exponent for the proportion of coded data. Default is 1.}

\item{exponent_available_data_points}{Exponent for the number of available data points. Default is 1.}

\item{exponent_lowest_language_score}{Exponent for the lowest language score. Default is 1.}

\item{exponent_taxonomic_diversity}{Exponent for the taxonomic diversity index. Default is 1.}
}
\value{
The optimal number of iterations with the highest quality score.
}
\description{
Function to define a quality score for each iteration as documented in the output created in F2 and identify the optimal number of iterations post-hoc. Exponents can be modified.
}
\examples{
# Assuming `documentation` is a data frame generated from `densify_steps` function
densify_score(documentation)

}
