% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F2_densify_steps.R
\name{densify_steps}
\alias{densify_steps}
\title{Iterative matrix densification according to specified criteria.}
\usage{
densify_steps(
  original_data,
  max_steps = 1,
  mean_type = "log_odds",
  taxonomy = F,
  original_register,
  tax_weight_factor = 0.99,
  coding_weight_factor = 0.99
)
}
\arguments{
\item{original_data}{A data frame with the glottocodes as row names and variable names as column names.
Any question marks, empty entries, or "NA"s must be represented as NAs.}

\item{max_steps}{An optional integer specifying the maximum number of iterations attempted during densification.
Default is 1.}

\item{mean_type}{A character string specifying the type of mean to be used for calculating the final weights.
Possible values are "arithmetic", "geometric", or "log_odds". Default is "log_odds".}

\item{taxonomy}{A logical indicating whether taxonomic diversity should be factored in for language pruning/retention.
If \code{TRUE}, taxonomic diversity is considered; if \code{FALSE}, it is not. Default is \code{FALSE}.}

\item{original_register}{A data frame containing the register retrieved from the \code{glottocode_taxonomy} function.
This parameter must be specified if \code{taxonomy = TRUE}.}

\item{tax_weight_factor}{A numeric value between 0 and 1 that determines the relative weight given to taxonomy in language pruning.
This parameter must be specified if \code{taxonomy = TRUE} and \code{mean_type = "log_odds"}.}

\item{coding_weight_factor}{A numeric value between 0 and 1 that determines the relative weight given to coding quality
(absolute coding score and weighted coding score) in language pruning. This parameter must be specified if \code{taxonomy = TRUE}
and \code{mean_type = "log_odds"}.}
}
\value{
A data frame with details about the matrix after each iteration of densification.
}
\description{
The output of this densification is a log-file, which specifies details about the matrix after
each iteration.
}
\examples{
# Assuming `original_data` and `original_register` are appropriate data frames
densify_steps(original_data, max_steps = 3, mean_type = "log_odds", taxonomy = TRUE, original_register, tax_weight_factor = 0.99, coding_weight_factor = 0.99)

}
