% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F1_glottocode_taxonomy.R
\name{glottocode_taxonomy}
\alias{glottocode_taxonomy}
\title{glottocode_taxonomy}
\usage{
glottocode_taxonomy(path_to_file)
}
\arguments{
\item{path_to_file}{The path to the input CSV file containing glottolog information.}
}
\value{
A data frame with the flattened glottolog tree for each languoid, including node levels.
}
\description{
Prepare glottolog info for all languages.
}
\details{
Our goal is to flatten out the glottolog tree for every languoid (glottocode) in the
database and represent the node as linear levels. Here, the top-level items
(roots of the glottolog tree) get mapped to node level 1 and the bottom-level items
(leaves) are mapped to the last node. Intermediate dummy nodes are created in between
and filled with NA if necessary.

For true singletons (leaves which are also roots), a virtual root is created during
the flattening process, e.g. Basque -> Basque
}
\examples{
# Assuming `path_to_file` is the path to the glottolog CSV file
glottocode_taxonomy(path_to_file)

}
\references{
Add references here (if available).
}
\seealso{
\code{\link{phytools}}, \code{\link{tidyverse}}, \code{\link{testthat}},
\code{\link{GetoptLong}}, \code{\link{readr}}
}
\concept{Linguistic Data Processing}
\keyword{data}
\keyword{flattening,}
\keyword{frame}
\keyword{glottolog,}
\keyword{leaves,}
\keyword{nodes,}
\keyword{taxonomy,}
